refactor(viewmodel): move out the types that were in the wrong layer
The last of #353. These were not merely in the wrong file — they were imported
by screens, components and widgets, which is a layering leak: every consumer
ended up depending on the one feature whose ViewModel happened to declare them.

PrayerTimeDisplay is the clearest case. It lived in HomeViewModel.kt and was
imported by eight files, including PrayerTimesViewModel — a ViewModel reaching
into an unrelated feature's ViewModel for a type it renders. It moves, with
DailyDua and List<PrayerTimeDisplay>.withClockState, to presentation/model/.

To domain/model/, because each is a fact about the content rather than about how
it is drawn: UnifiedBookmark and BookmarkType (they unify three content kinds and
carry the nav target each resolves to), UnifiedSearchResult (it belongs beside
LibrarySearchResults), DayPrayerTimes.

LocationCatalog.kt was not a ViewModel at all: 95 lines holding an enum, a
40-entry curated dataset and two list transforms. Those go to
domain/model/CityCatalog.kt with SearchLocation, and formatCoordinates — which is
UI formatting — goes to screens/settings/LocationFormat.kt, following the
screens/khatam/KhatamFormat.kt precedent.

Fixes the it.region!! that groupCitiesByRegion did after filtering for non-null:
a mapNotNull that cannot NPE, per the issue.

The acceptance box this closes is checkable, and now reads zero:

  grep -rn "import ...presentation.viewmodel" widget/ presentation/components/

Refs #353, #352
