+++ app/src/main/java/com/arshadshah/nimaz/presentation/screens/prayer/MonthlyPrayerTimesScreen.kt
@@ -86,7 +86,7 @@ import com.arshadshah.nimaz.presentation.theme.LocalUse24HourFormat
 import com.arshadshah.nimaz.presentation.theme.NimazColors
 import com.arshadshah.nimaz.presentation.theme.NimazCornerRadius
 import com.arshadshah.nimaz.presentation.theme.NimazSpacing
-import com.arshadshah.nimaz.presentation.viewmodel.prayer.DayPrayerTimes
+import com.arshadshah.nimaz.domain.model.DayPrayerTimes
 import com.arshadshah.nimaz.presentation.viewmodel.prayer.MonthlyPrayerTimesEvent
 import com.arshadshah.nimaz.presentation.viewmodel.prayer.MonthlyPrayerTimesViewModel
 import java.time.LocalDate
