+++ app/src/main/java/com/arshadshah/nimaz/presentation/viewmodel/home/HomeViewModel.kt
@@ -73,60 +73,9 @@ import kotlinx.coroutines.isActive
 import kotlinx.coroutines.launch
 import kotlinx.datetime.TimeZone
 import kotlinx.datetime.toLocalDateTime
-
-/**
- * A single dua surfaced on the home screen's "Today" section, picked to match
- * the current time of day (morning / evening / before sleep adhkar).
- */
-data class DailyDua(
-    val title: String,
-    val arabic: String,
-    val translation: String,
-    val source: String,
-    val categoryLabel: String,
-    val categoryIcon: String,
-)
-
-/**
- * One prayer row. Carries the prayer's **instant**, not a formatted string: the clock time is
- * rendered at the leaf so it honours `LocalUse24HourFormat` in the same frame the toggle flips,
- * and [isPassed]/[isCurrent]/[isNext] are re-derived at the leaf from the shared ticker via
- * [withClockState] rather than pushed once a second by a ViewModel loop.
- *
- * The three booleans default to `false`: a ViewModel publishes the facts (type, instant, status)
- * and the UI decides what they mean *now*.
- */
-data class PrayerTimeDisplay(
-    val type: PrayerType,
-    val name: String,
-    val timeAt: Instant,
-    val isPassed: Boolean = false,
-    val isCurrent: Boolean = false,
-    val isNext: Boolean = false,
-    val prayerStatus: PrayerStatus = PrayerStatus.NOT_PRAYED
-)
-
-/**
- * Re-derive the clock-dependent flags for [now]. Pure and cheap — call it from a composable that
- * reads `rememberNow()` so the list re-derives at the caller's tick resolution.
- *
- * Comparison is by instant, which fixes the long-standing bug where after Isha nothing highlighted
- * and before Fajr today's Isha rendered as "current" (see `core/util/PrayerClock.kt`).
- */
-fun List<PrayerTimeDisplay>.withClockState(now: Instant): List<PrayerTimeDisplay> {
-    if (isEmpty()) return this
-    val sorted = sortedBy { it.timeAt }
-    val instants = sorted.map { it.timeAt }
-    val nextIndex = nextPrayerIndexAt(instants, now)
-    val currentIndex = currentPrayerIndexAt(instants, now)
-    return sorted.mapIndexed { index, display ->
-        display.copy(
-            isPassed = display.timeAt <= now,
-            isCurrent = index == currentIndex,
-            isNext = index == nextIndex
-        )
-    }
-}
+import com.arshadshah.nimaz.presentation.model.DailyDua
+import com.arshadshah.nimaz.presentation.model.PrayerTimeDisplay
+import com.arshadshah.nimaz.presentation.model.withClockState
 
 @HiltViewModel
 class HomeViewModel @Inject constructor(
