+++ app/src/main/java/com/arshadshah/nimaz/presentation/viewmodel/location/LocationEvent.kt
@@ -1,5 +1,7 @@
 package com.arshadshah.nimaz.presentation.viewmodel.location
 
+import com.arshadshah.nimaz.domain.model.CityRegion
+import com.arshadshah.nimaz.domain.model.SearchLocation
 sealed interface LocationEvent {
     data class UpdateSearchQuery(val query: String) : LocationEvent
     data object Search : LocationEvent
