+++ app/src/main/java/com/arshadshah/nimaz/presentation/viewmodel/prayer/PrayerTimesUiState.kt
@@ -3,7 +3,7 @@ package com.arshadshah.nimaz.presentation.viewmodel.prayer
 import com.arshadshah.nimaz.domain.model.FallbackLocation
 import java.time.Instant
 import java.time.LocalDate
-import com.arshadshah.nimaz.presentation.viewmodel.home.PrayerTimeDisplay
+import com.arshadshah.nimaz.presentation.model.PrayerTimeDisplay
 
 /**
  * Drives the dedicated Prayer Times screen: a day pager over the prayer
