+++ app/src/main/java/com/arshadshah/nimaz/presentation/viewmodel/quran/BookmarksEvent.kt
@@ -1,5 +1,6 @@
 package com.arshadshah.nimaz.presentation.viewmodel.quran
 
+import com.arshadshah.nimaz.domain.model.BookmarkType
 sealed interface BookmarksEvent {
     data class SetFilter(val type: BookmarkType?) : BookmarksEvent
