+++ app/src/main/java/com/arshadshah/nimaz/presentation/viewmodel/quran/BookmarksUiState.kt
@@ -3,6 +3,8 @@ package com.arshadshah.nimaz.presentation.viewmodel.quran
 import com.arshadshah.nimaz.domain.model.DuaBookmark
 import com.arshadshah.nimaz.domain.model.HadithBookmark
 import com.arshadshah.nimaz.domain.model.QuranBookmark
+import com.arshadshah.nimaz.domain.model.BookmarkType
+import com.arshadshah.nimaz.domain.model.UnifiedBookmark
 
 data class BookmarksUiState(
     val allBookmarks: List<UnifiedBookmark> = emptyList(),
