+++ app/src/testDebug/java/com/arshadshah/nimaz/presentation/components/molecules/PrayerTimeCardTest.kt
@@ -6,7 +6,7 @@ import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
 import com.arshadshah.nimaz.domain.model.PrayerStatus
 import com.arshadshah.nimaz.domain.model.PrayerType
-import com.arshadshah.nimaz.presentation.viewmodel.home.PrayerTimeDisplay
+import com.arshadshah.nimaz.presentation.model.PrayerTimeDisplay
 import com.google.common.truth.Truth.assertThat
 import org.junit.Rule
 import org.junit.Test
